refactor: inline auth actions into authSlice, remove redundant actions.ts
- Move loginSuccess and logoutSuccess from extraReducers/authSlice
  into the reducers block of authSlice (idiomatic Redux Toolkit)
- Delete src/store/actions.ts (no longer needed)
- Remove unused AUTH_LOGIN_SUCCESS/AUTH_LOGOUT_SUCCESS constants from src/store/types.ts
- Update all files importing from '.../store/actions' to import from '.../store/authSlice'
- Action type strings remain identical (auth/loginSuccess, auth/logoutSuccess)
