+++ src/store/types.ts
@@ -1,10 +1,5 @@
 // This file contains all the action types for the redux store.
 
-// Auth Action Types
-export const AUTH_LOGIN_SUCCESS = 'auth/loginSuccess';
-export const AUTH_LOGOUT_SUCCESS = 'auth/logoutSuccess';
-
-
 // Represents the shape of a user object
 export interface User {
   id: number;
