+++ src/tests/components/game/GameLeaderboard.scroll.test.tsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import i18n from '../../../locale/i18n';
 import store from '../../../store';
 import GameLeaderboard from '../../../components/game/GameLeaderboard';
-import { loginSuccess, logoutSuccess } from '../../../store/actions';
+import { loginSuccess, logoutSuccess } from "../../../store/authSlice";
 import { http, HttpResponse } from 'msw';
 import { server } from '../../mocks/server';
 import { API_ENDPOINTS } from '../../../services/apiEndpoints';
