+++ src/tests/components/game/GameLeaderboard.test.tsx
@@ -5,7 +5,7 @@ import { Provider } from 'react-redux';
 import i18n from '../../../locale/i18n';
 import store from '../../../store';
 import GameLeaderboard from '../../../components/game/GameLeaderboard';
-import { loginSuccess, logoutSuccess } from '../../../store/actions';
+import { loginSuccess, logoutSuccess } from "../../../store/authSlice";
 
 const defaultAdminFields = {
   logins_remaining_for_staff: 0,
